refactor: 핀 모델 파일 위치 변경
