+++ app/src/main/java/com/issueissyu/fe/domain/model/MapPinModels.kt
@@ -1,5 +1,8 @@
 package com.issueissyu.fe.domain.model
 
+import com.issueissyu.fe.domain.model.pin.PinCategory
+import com.issueissyu.fe.domain.model.pin.PinCoordinate
+
 data class MapPinMarker(
     val pinId: String,
     val category: PinCategory,
