+++ app/src/main/java/com/issueissyu/fe/ui/components/map/MapCoordinateMapper.kt
@@ -1,6 +1,6 @@
 package com.issueissyu.fe.ui.components.map
 
-import com.issueissyu.fe.domain.model.PinCoordinate
+import com.issueissyu.fe.domain.model.pin.PinCoordinate
 import com.naver.maps.geometry.LatLng
 
 fun PinCoordinate.toLatLng(): LatLng {
