+++ app/src/main/java/com/issueissyu/fe/ui/screens/patchnote/PatchNoteViewModel.kt
@@ -2,7 +2,7 @@ package com.issueissyu.fe.ui.screens.patchnote
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.issueissyu.fe.domain.model.IssuePinDetail
+import com.issueissyu.fe.domain.model.pin.IssuePinDetail
 import com.issueissyu.fe.domain.repository.PinRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
